requirement.txt
asgiref==3.3.1
Django==3.1.4
django-crontab==0.7.1
djangorestframework==3.12.2
feedparser==6.0.2
pytz==2020.4
sgmllib3k==1.0.0
sqlparse==0.4.1
